# Conflicts:
#	crest/Assets/Crest/Crest/Scripts/LodData/RegisterFlowInput.cs

+++ crest/Assets/Crest/Crest/Shaders/OceanEmission.hlsl
@@ -101,7 +101,7 @@ void ApplyCaustics(in const half3 i_view, in const half3 i_lightDir, in const fl
 
 	const float3 scenePosUV = WorldToUV(scenePos.xz, cascadeData1, _LD_SliceIndex + 1);
 
-	half3 disp = 0.;
+	float3 disp = 0.0;
 	// this gives height at displaced position, not exactly at query position.. but it helps. i cant pass this from vert shader
 	// because i dont know it at scene pos.
 	SampleDisplacements(_LD_TexArray_AnimatedWaves, scenePosUV, 1.0, disp);
